Tons of work on making it sane python.

+++ src/process_manager.py
@@ -18,42 +18,42 @@ import subprocess
 import abc
 
 class ProcessManager(object):
-	""" ProcessManager is a manager for a process executed by the subprocess module.
+    """ ProcessManager is a manager for a process executed by the subprocess module.
 
-	It exists so that we can have a common interface for checking the status of an executed process-
-	no need to muck around with subprocess polling methods, ProcessManager wraps these for ease of use.
+    It exists so that we can have a common interface for checking the status of an executed process-
+    no need to muck around with subprocess polling methods, ProcessManager wraps these for ease of use.
 
-	The method get_command(self) must be implemented in subclasses.
-	"""
+    The method get_command(self) must be implemented in subclasses.
+    """
 
-	__metaclass__ = abc.ABCMeta
+    __metaclass__ = abc.ABCMeta
 
-	def __init__(self):
-		self.process_started = False
-		self.process_instance = None
+    def __init__(self):
+        self.process_started = False
+        self.process_instance = None
 
-	def start(self):
-		self.process_started = True
-		self.process_instance = subprocess.Popen(get_command())
+    def start(self):
+        self.process_started = True
+        self.process_instance = subprocess.Popen(self.get_command())
 
-	def stop(self, wait=False):
-		if self.process_instance is not None:
-			self.process_instance.terminate()
-			if wait: self.process_instance.wait()
-			self.process_started = False
+    def stop(self, wait=True):
+        if self.process_instance is not None and self.is_running():
+            self.process_instance.terminate()
+            if wait: self.process_instance.wait()
+        self.process_started = False
 
-	def is_started(self):
-		""" Returns whether or not the process has been started.
-			Resets after process termination. """
-		return self.process_started
+    def is_started(self):
+        """ Returns whether or not the process has been started.
+            Resets after process termination. """
+        return self.process_started
 
-	def is_running(self):
-		""" Returns whether or not the process is currently running. """
-    	if self.process_instance is None:
-        	return False
-   		else:
-        	return self.process_instance.poll() is None
+    def is_running(self):
+        """ Returns whether or not the process is currently running. """
+        if self.process_instance is None:
+            return False
+        else:
+            return self.process_instance.poll() is None
 
-	@abc.abstractmethod
-	def get_command(self):
-		return
+    @abc.abstractmethod
+    def get_command(self):
+        return
